+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 # Changelog
 
+## [0.2.0] - 2024-02-01
+
+### Breaking changes
+
+- `TypedUuidTag`s are now required to be valid ASCII identifiers, with hyphens also supported.
+
+### Misc
+
+- Added `#[forbid(unsafe_code)]`. Thanks [Robert Lynch](https://github.com/rob0rt) for the contribution!
+
 ## [0.1.0] - 2024-01-30
 
 Initial release.
 
+[0.2.0]: https://github.com/oxidecomputer/newtype-uuid/releases/newtype-uuid-0.2.0
 [0.1.0]: https://github.com/oxidecomputer/newtype-uuid/releases/newtype-uuid-0.1.0